Rename common parameter `path` to `uri` when from `req.uri()`.
There is a subtle distinction between `route.uri` and `req.uri()`
in that the former always contains the matched URI pattern,
and not the actual URI received when that URI contains dynamic parts.

That distinction is due to the on_request side of the fairing happening
before the request has been matched to a route. Because of that, there
is no value for `route.uri` at request time / logging.

For example, suppose there is a route of path `/items/<id>` and a
request has been received with the actual URI of `/items/123`.
The `route.uri` would equal "/items/<id>",
and `req.uri()` would equal "/items/123".

From now on, the log parameter `path` will contain the `route.uri`,
and a new parameter of `uri` will contain the actual URI from `req.uri()`.